projects/cv_profile_photo/main.py
# Import libraries
import numpy as np
import stylegan2.dnnlib as dnnlib
import stylegan2.dnnlib.tflib as tflib
import PIL.Image
import cv2
import os
os.environ['PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION'] = 'python'

# Initialize TensorFlow session
tflib.init_tf()

# Load pre-trained network
url = 'https://nvlabs-fi-cdn.nvidia.com/stylegan2-ada/pretrained/ffhq.pkl'
with dnnlib.util.open_url(url) as f:
    _G, _D, Gs = pickle.load(f)

# Define image processing functions
def convert_to_pil(image):
    # Convert a NumPy array to a PIL image
    image = np.clip(np.rint((image + 1) * 127.5), 0, 255).astype(np.uint8)
    image = image.transpose(1, 2, 0)
    image = PIL.Image.fromarray(image, 'RGB')
    return image

def align_face(image):
    # Align a face image using OpenCV
    # You can use any face detection and alignment method you prefer
    # Here we use a simple Haar cascade classifier
    face_cascade = cv2.CascadeClassifier('haarcascade_frontalface_default.xml')
    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    faces = face_cascade.detectMultiScale(gray, 1.3, 5)
    if len(faces) == 0:
        return image
    else:
        # Crop the largest face detected
        x, y, w, h = max(faces, key=lambda x: x[2] * x[3])
        image = image[y:y+h, x:x+w]
        # Resize the image to 256x256
        image = cv2.resize(image, (256, 256))
        return image

def change_background(image, color=(255, 255, 255)):
    # Change the background color of a face image
    # You can use any background removal and replacement method you prefer
    # Here we use a simple thresholding and masking technique
    # Convert the image to HSV color space
    hsv = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
    # Define the lower and upper bounds of skin color
    lower = np.array([0, 48, 80], dtype=np.uint8)
    upper = np.array([20, 255, 255], dtype=np.uint8)
    # Create a mask for skin color
    mask = cv2.inRange(hsv, lower, upper)
    # Apply the mask to the image
    image = cv2.bitwise_and(image, image, mask=mask)
    # Create a background image with the desired color
    background = np.full(image.shape, color, dtype=np.uint8)
    # Combine the image and the background
    image = cv2.bitwise_or(image, background)
    return image

# Load your profile pic
profile_pic = PIL.Image.open('your_profile_pic.jpg')
# Convert it to a NumPy array
profile_pic = np.array(profile_pic)

# Align your face
aligned_face = align_face(profile_pic)

# Generate a latent vector from your aligned face
# This is the input for the StyleGAN2 model
# You can use any latent vector optimization method you prefer
# Here we use a simple gradient descent approach
# Define the loss function
def loss_fun(latent_vector):
    # Generate a synthetic image from the latent vector
    synth_image = Gs.components.synthesis.run(latent_vector, randomize_noise=False, minibatch_size=1)
    synth_image = convert_to_pil(synth_image[0])
    synth_image = np.array(synth_image)
    # Compute the mean squared error between the synthetic image and the aligned face
    mse = np.mean((synth_image - aligned_face) ** 2)
    return mse

# Initialize a random latent vector
latent_vector = np.random.randn(1, Gs.input_shape[1])
# Define the optimizer
optimizer = tf.optimizers.Adam(learning_rate=0.01)
# Define the number of iterations
iterations = 100
# Optimize the latent vector
for i in range(iterations):
    # Compute the gradients of the loss function with respect to the latent vector
    with tf.GradientTape() as tape:
        tape.watch(latent_vector)
        loss = loss_fun(latent_vector)
    # Update the latent vector using the optimizer
    gradients = tape.gradient(loss, latent_vector)
    optimizer.apply_gradients(zip([gradients], [latent_vector]))
    # Print the progress
    print(f'Iteration {i+1}: Loss = {loss}')

# Generate a CV photo from the optimized latent vector
cv_photo = Gs.components.synthesis.run(latent_vector, randomize_noise=False, minibatch_size=1)
cv_photo = convert_to_pil(cv_photo[0])
cv_photo = np.array(cv_photo)

# Change the background color of the CV photo
cv_photo = change_background(cv_photo, color=(255, 255, 255))

# Save the CV photo
cv2.imwrite('your_cv_photo.jpg', cv_photo)