projects/cv_profile_photo/requirementx.txt
numpy==1.21.2
dnnlib==0.0.0
tensorflow==2.6.0
Pillow==8.3.2
opencv-python==4.5.3.56
